refactor(zql): move some common code to `util`, add tests
